Let each module claim its own timers, instead of the poller reaching in
The previous commit put both timer claimers in `schedule` and imported
`workflow` and `agent` models to do it. The architecture ratchet called that
correctly: `schedule -> workflow` is a cycle, because workflow already depends on
schedule, and `agent -> schedule` is forbidden outright.

The boundary was in the wrong place, not the gate. A module that owns a table
should own the query against it, so the claimers now live in `workflow` and
`agent`, and the poller takes them as injected callables. Composition wires them
together in `worker_lifespan`, which is where crossing module boundaries is the
job.

`ClaimedTimer` and the lease helpers moved to `app/core/domain/timers.py` rather
than staying in `schedule`, for the same reason: `agent` may not import
`schedule`, and a claimer forced across a module boundary to name its own return
type is a boundary in the wrong place too.

No behaviour change -- same claims, same leases, same payloads, same five tests.
