+++ lemma-backend/app/core/infrastructure/jobs/streaq_runtime.py
@@ -548,11 +548,17 @@ async def worker_lifespan() -> AsyncGenerator[AppWorkerContext]:
     # Fires due schedules and timers. Runs on every worker replica: the poll
     # claims with FOR UPDATE SKIP LOCKED, so replicas share the work rather than
     # duplicating it, and there is no leader to lose.
+    from app.modules.agent.services.due_snooze_claimer import claim_due_snooze_waits
     from app.modules.schedule.services.schedule_poller import run_schedule_poller
+    from app.modules.workflow.services.due_wait_claimer import (
+        claim_due_workflow_waits,
+    )
 
     schedule_poller_task = create_background_task(
         run_schedule_poller(
             context.uow_factory,
+            # Injected here, where crossing module boundaries is the job.
+            timer_claimers=(claim_due_workflow_waits, claim_due_snooze_waits),
             interval_seconds=settings.schedule_poll_interval_seconds,
         ),
         name="schedule-poller",
